RangeChart.jsx changed to use react-chartjs Implementing momentjs to format data visualization
